refactor(datalayer): modularize schema, extract load pipeline, clarify layers
Split schema into one module per table, move fetch/normalize/store
orchestration into load.py, and document strict layer contracts so the
datalayer is cleaner to extend without changing the public query API.
