+++ datalayer/docs/02_surfacing_names.md
@@ -1,3 +1,5 @@
+> **Note:** Historical design doc. Paths like `ddl.py`, `defaults.py`, or monolith `schema/models.py` may be stale. See `datalayer/AGENTS.md` and `docs/architecture.md` for the current layout (one module per table under `schema/`, queries split by domain, load orchestration in `load.py`).
+
 ---
 
 # Design Doc: Name-Enriched Outputs + Name-Based Query Inputs
