+++ datalayer/sleeper_data/__init__.py
@@ -1,15 +1,14 @@
-"""Public package exports for sleeper data layer."""
+"""Public package exports for sleeper data layer.
+
+Prefer ``SleeperLeagueData`` as the integration surface. Store helpers and
+schema internals are available via deeper imports when needed by tests.
+"""
 
 from .config import SleeperConfig, load_config
-from .schema import models as schema_models
 from .sleeper_league_data import SleeperLeagueData
-from .store.sqlite_store import bulk_insert, create_tables
 
 __all__ = [
     "SleeperConfig",
     "SleeperLeagueData",
     "load_config",
-    "schema_models",
-    "bulk_insert",
-    "create_tables",
 ]
