+++ datalayer/sleeper_data/normalize/transactions.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import json
 from typing import Any, Iterable, Mapping
 
-from ..schema.models import Transaction, TransactionMove
+from ..schema import Transaction, TransactionMove
 
 
 def _json_dumps(value: Any) -> str | None:
