+++ datalayer/sleeper_data/normalize/users.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import json
 from typing import Any, Iterable, Mapping
 
-from ..schema.models import User
+from ..schema import User
 
 
 def _json_dumps(value: Any) -> str | None:
