+++ datalayer/sleeper_data/sleeper_league_data.py
@@ -1,47 +1,25 @@
-"""Facade for loading Sleeper league data into SQLite."""
+"""Public facade for Sleeper league data: load orchestration + curated queries.
+
+Layer contract:
+- Owns engine/connection lifecycle and the public query API.
+- Delegates fetch/normalize/store to ``load.load_league``.
+- Delegates SQL to ``queries.*``; does not contain business SQL itself.
+- Tools/CLI should call this class only.
+"""
 
 from __future__ import annotations
 
 import os
 import sqlite3
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Mapping, Optional
 
 from sqlalchemy import create_engine, text
+from sqlalchemy.engine import Connection
 
 from .config import SleeperConfig, load_config
-from .normalize import (
-    apply_traded_picks,
-    derive_games,
-    derive_team_profiles,
-    normalize_bracket,
-    normalize_league,
-    normalize_matchups,
-    normalize_players,
-    normalize_roster_players,
-    normalize_rosters,
-    normalize_standings,
-    normalize_transaction_moves,
-    normalize_transactions,
-    normalize_users,
-    seed_draft_picks,
-)
-from .schema.models import SeasonContext, StandingsWeek
-from .sleeper_api import (
-    SleeperClient,
-    get_league,
-    get_league_rosters,
-    get_league_users,
-    get_losers_bracket,
-    get_matchups,
-    get_players,
-    get_state,
-    get_traded_picks,
-    get_transactions as api_get_transactions,
-    get_winners_bracket,
-)
-from .store.sqlite_store import bulk_insert, create_tables
+from .load import load_league
+from .sleeper_api import SleeperClient
 from .queries import (
     get_bench_analysis,
     get_league_snapshot,
@@ -79,8 +57,13 @@ class SleeperLeagueData:
         self.week_override = resolved_config.week_override
         self.client = client or SleeperClient()
         self.engine = None
-        self._query_conn = None
+        self._query_conn: Connection | None = None
         self.effective_week: Optional[int] = None
+
+    def _conn(self) -> Connection:
+        if not self._query_conn:
+            raise RuntimeError("Data not loaded. Call load() before querying.")
+        return self._query_conn
 
     @classmethod
     def from_file(cls, path: str | os.PathLike[str]) -> "SleeperLeagueData":
@@ -136,240 +119,13 @@ class SleeperLeagueData:
         self.engine = create_engine(
             "sqlite://", connect_args={"check_same_thread": False}
         )
-
-        raw_league = get_league(self.league_id, client=self.client)
-        raw_users = get_league_users(self.league_id, client=self.client)
-        raw_rosters = get_league_rosters(self.league_id, client=self.client)
-        raw_state = get_state("nfl", client=self.client)
-
-        league = normalize_league(raw_league)
-        users = normalize_users(raw_users)
-        rosters = normalize_rosters(raw_rosters, league_id=self.league_id)
-        roster_players = normalize_roster_players(raw_rosters, league_id=self.league_id)
-        team_profiles = derive_team_profiles(
-            raw_rosters, raw_users, league_id=self.league_id
+        result = load_league(
+            self.engine,
+            client=self.client,
+            league_id=self.league_id,
+            week_override=self.week_override,
         )
-
-        with self.engine.begin() as conn:
-            create_tables(conn)
-
-            bulk_insert(conn, league.table_name, [league])
-            if users:
-                bulk_insert(conn, users[0].table_name, users)
-            if rosters:
-                bulk_insert(conn, rosters[0].table_name, rosters)
-            if team_profiles:
-                bulk_insert(conn, team_profiles[0].table_name, team_profiles)
-
-            draft_rounds = int(
-                (raw_league.get("settings") or {}).get("draft_rounds") or 0
-            )
-            draft_picks = seed_draft_picks(
-                rosters, self.league_id, league.season, draft_rounds
-            )
-            if draft_picks:
-                bulk_insert(conn, draft_picks[0].table_name, draft_picks)
-
-            raw_traded_picks = get_traded_picks(self.league_id, client=self.client)
-            apply_traded_picks(conn, raw_traded_picks, self.league_id)
-
-            raw_players = get_players("nfl", client=self.client)
-            players = normalize_players(raw_players)
-            if players:
-                bulk_insert(conn, players[0].table_name, players)
-            if roster_players:
-                bulk_insert(conn, roster_players[0].table_name, roster_players)
-
-            computed_week = int(raw_state.get("week") or 0)
-            effective_week = int(self.week_override or computed_week or 0)
-            season = str(raw_league.get("season") or raw_state.get("season") or "")
-            league_average_match = (
-                int(
-                    (raw_league.get("settings") or {}).get("league_average_match") or 0
-                )
-                if raw_league.get("settings") is not None
-                else 0
-            )
-            chars_per_week = 2 if league_average_match == 1 else 1
-
-            playoff_week_start = (
-                int(league.playoff_week_start)
-                if league.playoff_week_start is not None
-                else None
-            )
-
-            def _record_string_to_weeks(
-                record_string: str | None,
-            ) -> list[tuple[int, int, int, int]]:
-                if not record_string:
-                    return []
-                trimmed = "".join(
-                    ch for ch in record_string.strip().upper() if ch.strip()
-                )
-                if not trimmed:
-                    return []
-                week_count = len(trimmed) // chars_per_week
-                results: list[tuple[int, int, int, int]] = []
-                wins = 0
-                losses = 0
-                ties = 0
-                for week in range(1, week_count + 1):
-                    end_idx = week * chars_per_week
-                    slice_value = trimmed[end_idx - chars_per_week : end_idx]
-                    for outcome in slice_value:
-                        if outcome == "W":
-                            wins += 1
-                        elif outcome == "L":
-                            losses += 1
-                        elif outcome == "T":
-                            ties += 1
-                    results.append((week, wins, losses, ties))
-                return results
-
-            self.effective_week = effective_week
-            if effective_week > 0:
-                for week in range(1, effective_week + 1):
-                    raw_matchups = get_matchups(
-                        self.league_id, week, client=self.client
-                    )
-                    matchup_rows, player_performances = normalize_matchups(
-                        raw_matchups,
-                        league_id=self.league_id,
-                        season=season,
-                        week=week,
-                    )
-                    is_playoffs = playoff_week_start is not None and week >= int(
-                        playoff_week_start
-                    )
-                    games = derive_games(matchup_rows, is_playoffs=is_playoffs)
-                    if matchup_rows:
-                        bulk_insert(conn, matchup_rows[0].table_name, matchup_rows)
-                    if player_performances:
-                        bulk_insert(
-                            conn,
-                            player_performances[0].table_name,
-                            player_performances,
-                        )
-                    if games:
-                        bulk_insert(conn, games[0].table_name, games)
-
-                    raw_transactions = api_get_transactions(
-                        self.league_id, week, client=self.client
-                    )
-                    transactions = normalize_transactions(
-                        raw_transactions,
-                        league_id=self.league_id,
-                        season=season,
-                        week=week,
-                    )
-                    moves = normalize_transaction_moves(raw_transactions)
-                    if transactions:
-                        bulk_insert(conn, transactions[0].table_name, transactions)
-                    if moves:
-                        bulk_insert(conn, moves[0].table_name, moves)
-
-                record_standings: list[StandingsWeek] = []
-                record_weeks: set[int] = set()
-
-                for raw_roster in raw_rosters:
-                    roster_id = int(raw_roster["roster_id"])
-                    record_string = (raw_roster.get("metadata") or {}).get("record")
-                    if isinstance(record_string, list):
-                        record_string = "".join(
-                            str(item) for item in record_string if item
-                        )
-                    if isinstance(record_string, str):
-                        for week, wins, losses, ties in _record_string_to_weeks(
-                            record_string
-                        ):
-                            if week > effective_week:
-                                break
-                            if (
-                                playoff_week_start is not None
-                                and week >= playoff_week_start
-                            ):
-                                continue
-                            record_standings.append(
-                                StandingsWeek(
-                                    league_id=self.league_id,
-                                    season=season,
-                                    week=int(week),
-                                    roster_id=roster_id,
-                                    wins=wins,
-                                    losses=losses,
-                                    ties=ties,
-                                    points_for=0.0,
-                                    points_against=0.0,
-                                    rank=None,
-                                    streak_type=None,
-                                    streak_len=None,
-                                )
-                            )
-                            record_weeks.add(int(week))
-
-                if record_standings:
-                    bulk_insert(
-                        conn,
-                        record_standings[0].table_name,
-                        record_standings,
-                    )
-
-                should_insert_current = False
-                if not record_weeks:
-                    should_insert_current = True
-                elif effective_week > max(record_weeks):
-                    if (
-                        playoff_week_start is None
-                        or effective_week < playoff_week_start
-                    ):
-                        should_insert_current = True
-
-                if should_insert_current:
-                    standings = normalize_standings(
-                        raw_rosters,
-                        league_id=self.league_id,
-                        season=season,
-                        week=effective_week,
-                    )
-                    if standings:
-                        bulk_insert(conn, standings[0].table_name, standings)
-
-            # Only load playoff brackets if we've reached the playoff weeks.
-            # Loading them earlier would leak future results when using
-            # week_override to view a mid-season snapshot.
-            should_load_brackets = (
-                playoff_week_start is None
-                or effective_week >= playoff_week_start
-            )
-            if should_load_brackets:
-                raw_winners = get_winners_bracket(self.league_id, client=self.client)
-                raw_losers = get_losers_bracket(self.league_id, client=self.client)
-                winners = normalize_bracket(
-                    raw_winners,
-                    league_id=self.league_id,
-                    season=season,
-                    bracket_type="winners",
-                )
-                losers = normalize_bracket(
-                    raw_losers,
-                    league_id=self.league_id,
-                    season=season,
-                    bracket_type="losers",
-                )
-                if winners:
-                    bulk_insert(conn, winners[0].table_name, winners)
-                if losers:
-                    bulk_insert(conn, losers[0].table_name, losers)
-
-            season_context = SeasonContext(
-                league_id=self.league_id,
-                computed_week=computed_week,
-                override_week=self.week_override,
-                effective_week=effective_week,
-                generated_at=datetime.now(timezone.utc).isoformat(),
-            )
-            bulk_insert(conn, season_context.table_name, [season_context])
-
+        self.effective_week = result.effective_week
         # Open a long-lived connection for queries
         self._query_conn = self.engine.connect()
 
@@ -392,14 +148,24 @@ class SleeperLeagueData:
 
         return output_path
 
+    def _get_effective_week(self, week: int | None = None) -> int | None:
+        """Get effective week, defaulting to current week if not specified."""
+        if week is not None:
+            return week
+        conn = self._query_conn
+        if conn is None:
+            return None
+        result = conn.execute(
+            text("SELECT effective_week FROM season_context LIMIT 1")
+        ).first()
+        return result[0] if result else None
+
     def get_league_snapshot(self, week: int | None = None) -> dict[str, Any]:
         """Get league standings, games, and transactions for a week.
 
-        See queries.defaults.get_league_snapshot for full return structure.
+        See queries.league.get_league_snapshot for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_league_snapshot(self._query_conn, week)
+        return get_league_snapshot(self._conn(), week)
 
     def get_bench_analysis(
         self, roster_key: Any = None, week: int | None = None
@@ -413,15 +179,14 @@ class SleeperLeagueData:
 
         See queries.league.get_bench_analysis for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             if roster_key is not None:
                 return {"found": False, "roster_key": roster_key}
             return []
         return get_bench_analysis(
-            self._query_conn, self.league_id, int(effective_week), roster_key
+            conn, self.league_id, int(effective_week), roster_key
         )
 
     def get_standings(self, week: int | None = None) -> dict[str, Any]:
@@ -432,9 +197,7 @@ class SleeperLeagueData:
 
         See queries.league.get_standings for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_standings(self._query_conn, self.league_id, week)
+        return get_standings(self._conn(), self.league_id, week)
 
     def get_team_dossier(
         self, roster_key: Any, week: int | None = None
@@ -445,11 +208,9 @@ class SleeperLeagueData:
             roster_key: Team name, manager name, or roster_id.
             week: Week for standings (defaults to current week).
 
-        See queries.defaults.get_team_dossier for full return structure.
+        See queries.team.get_team_dossier for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_team_dossier(self._query_conn, self.league_id, roster_key, week)
+        return get_team_dossier(self._conn(), self.league_id, roster_key, week)
 
     def get_team_schedule(self, roster_key: Any) -> dict[str, Any]:
         """Get full season schedule with game-by-game results.
@@ -457,22 +218,9 @@ class SleeperLeagueData:
         Args:
             roster_key: Team name, manager name, or roster_id.
 
-        See queries.defaults.get_team_schedule for full return structure.
+        See queries.team.get_team_schedule for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_team_schedule(self._query_conn, self.league_id, roster_key)
-
-    def _get_effective_week(self, week: int | None = None) -> int | None:
-        """Get effective week, defaulting to current week if not specified."""
-        if week is not None:
-            return week
-        if not self._query_conn:
-            return None
-        result = self._query_conn.execute(
-            text("SELECT effective_week FROM season_context LIMIT 1")
-        ).first()
-        return result[0] if result else None
+        return get_team_schedule(self._conn(), self.league_id, roster_key)
 
     def get_week_games(self, week: int | None = None) -> list[dict[str, Any]]:
         """Get all matchup games for a week with scores and winners.
@@ -482,12 +230,11 @@ class SleeperLeagueData:
 
         See queries.league.get_week_games for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return []
-        return get_week_games(self._query_conn, self.league_id, int(effective_week))
+        return get_week_games(conn, self.league_id, int(effective_week))
 
     def get_week_games_with_players(
         self, week: int | None = None
@@ -499,13 +246,12 @@ class SleeperLeagueData:
 
         See queries.league.get_week_games_with_players for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return []
         return get_week_games_with_players(
-            self._query_conn, self.league_id, int(effective_week)
+            conn, self.league_id, int(effective_week)
         )
 
     def get_team_game(self, roster_key: Any, week: int | None = None) -> dict[str, Any]:
@@ -517,12 +263,11 @@ class SleeperLeagueData:
 
         See queries.league.get_team_game for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return {"found": False, "roster_key": roster_key}
-        return get_team_game(self._query_conn, self.league_id, int(effective_week), roster_key)
+        return get_team_game(conn, self.league_id, int(effective_week), roster_key)
 
     def get_team_game_with_players(
         self, roster_key: Any, week: int | None = None
@@ -535,13 +280,12 @@ class SleeperLeagueData:
 
         See queries.league.get_team_game_with_players for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return {"found": False, "roster_key": roster_key}
         return get_team_game_with_players(
-            self._query_conn, self.league_id, int(effective_week), roster_key
+            conn, self.league_id, int(effective_week), roster_key
         )
 
     def get_week_player_leaderboard(
@@ -555,13 +299,12 @@ class SleeperLeagueData:
 
         See queries.league.get_week_player_leaderboard for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return []
         return get_week_player_leaderboard(
-            self._query_conn, self.league_id, int(effective_week), limit=limit
+            conn, self.league_id, int(effective_week), limit=limit
         )
 
     def get_season_leaders(
@@ -588,10 +331,8 @@ class SleeperLeagueData:
 
         See queries.league.get_season_leaders for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
         return get_season_leaders(
-            self._query_conn,
+            self._conn(),
             self.league_id,
             week_from=week_from,
             week_to=week_to,
@@ -611,9 +352,9 @@ class SleeperLeagueData:
 
         See queries.transactions.get_transactions for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return query_get_transactions(self._query_conn, self.league_id, week_from, week_to)
+        return query_get_transactions(
+            self._conn(), self.league_id, week_from, week_to
+        )
 
     def get_team_transactions(
         self, roster_key: Any, week_from: int, week_to: int
@@ -627,10 +368,8 @@ class SleeperLeagueData:
 
         See queries.transactions.get_team_transactions for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
         return get_team_transactions(
-            self._query_conn, self.league_id, week_from, week_to, roster_key
+            self._conn(), self.league_id, week_from, week_to, roster_key
         )
 
     def get_week_transactions(self, week: int | None = None) -> list[dict[str, Any]]:
@@ -641,13 +380,12 @@ class SleeperLeagueData:
 
         See queries.transactions.get_transactions for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         effective_week = self._get_effective_week(week)
         if effective_week is None:
             return []
         return query_get_transactions(
-            self._query_conn, self.league_id, effective_week, effective_week
+            conn, self.league_id, effective_week, effective_week
         )
 
     def get_team_week_transactions(
@@ -665,8 +403,7 @@ class SleeperLeagueData:
 
         See queries.transactions.get_team_transactions for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         if week_from is not None:
             resolved_from = week_from
             resolved_to = week_to if week_to is not None else week_from
@@ -677,7 +414,7 @@ class SleeperLeagueData:
             resolved_from = effective_week
             resolved_to = week_to if week_to is not None else effective_week
         return get_team_transactions(
-            self._query_conn, self.league_id, resolved_from, resolved_to, roster_key
+            conn, self.league_id, resolved_from, resolved_to, roster_key
         )
 
     def get_player_summary(self, player_key: Any) -> dict[str, Any]:
@@ -686,11 +423,9 @@ class SleeperLeagueData:
         Args:
             player_key: Player name or player_id.
 
-        See queries.defaults.get_player_summary for full return structure.
+        See queries.player.get_player_summary for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_player_summary(self._query_conn, player_key)
+        return get_player_summary(self._conn(), player_key)
 
     def get_player_weekly_log(
         self,
@@ -707,11 +442,12 @@ class SleeperLeagueData:
 
         See queries.player.get_player_weekly_log for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
         return get_player_weekly_log(
-            self._query_conn, self.league_id, player_key,
-            week_from=week_from, week_to=week_to,
+            self._conn(),
+            self.league_id,
+            player_key,
+            week_from=week_from,
+            week_to=week_to,
         )
 
     def get_roster_current(self, roster_key: Any) -> dict[str, Any]:
@@ -724,15 +460,14 @@ class SleeperLeagueData:
         Args:
             roster_key: Team name, manager name, or roster_id.
 
-        See queries.defaults.get_roster_current for full return structure.
+        See queries.team.get_roster_current for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
+        conn = self._conn()
         if self.week_override is not None and self.effective_week is not None:
             return get_roster_snapshot(
-                self._query_conn, self.league_id, roster_key, self.effective_week
+                conn, self.league_id, roster_key, self.effective_week
             )
-        return get_roster_current(self._query_conn, self.league_id, roster_key)
+        return get_roster_current(conn, self.league_id, roster_key)
 
     def get_roster_snapshot(self, roster_key: Any, week: int) -> dict[str, Any]:
         """Get a team's roster as it was during a specific week.
@@ -741,11 +476,11 @@ class SleeperLeagueData:
             roster_key: Team name, manager name, or roster_id.
             week: The week number to query.
 
-        See queries.defaults.get_roster_snapshot for full return structure.
+        See queries.team.get_roster_snapshot for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return get_roster_snapshot(self._query_conn, self.league_id, roster_key, week)
+        return get_roster_snapshot(
+            self._conn(), self.league_id, roster_key, week
+        )
 
     def run_sql(
         self,
@@ -763,9 +498,7 @@ class SleeperLeagueData:
 
         See queries.sql_tool.run_sql for full documentation and table list.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return run_sql(self._query_conn, query, params, limit=limit)
+        return run_sql(self._conn(), query, params, limit=limit)
 
     def get_playoff_bracket(self, bracket_type: str | None = None) -> dict[str, Any]:
         """Get the playoff bracket structure with team names and results.
@@ -775,9 +508,9 @@ class SleeperLeagueData:
 
         See queries.playoffs.get_playoff_bracket for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return query_get_playoff_bracket(self._query_conn, self.league_id, bracket_type)
+        return query_get_playoff_bracket(
+            self._conn(), self.league_id, bracket_type
+        )
 
     def get_team_playoff_path(self, roster_key: Any) -> dict[str, Any]:
         """Get a specific team's playoff bracket journey.
@@ -787,6 +520,6 @@ class SleeperLeagueData:
 
         See queries.playoffs.get_team_playoff_path for full return structure.
         """
-        if not self._query_conn:
-            raise RuntimeError("Data not loaded. Call load() before querying.")
-        return query_get_team_playoff_path(self._query_conn, self.league_id, roster_key)
+        return query_get_team_playoff_path(
+            self._conn(), self.league_id, roster_key
+        )
